backend/requirements.txt
uvicorn 
smolagents
smolagents[openai]
smolagents[gradio]
python-multipart
dotenv
sqlalchemy
fastapi
gradio
transformers
psycopg[binary]
psycopg2-binary
python-jose[cryptography]
passlib[argon2]
python-dotenv
email-validator
PyJWT
langgraph
lanchain
langchain-openai
langgraph-checkpoint-postgres
